Remove an IFD when importing nixos-appstream-data.
I recently added a flake for nixos-appstream-data. This allows us to
depend on that flake instead of doing an IFD to load its
default.nix.

An IFD in nix-software-center is problematic for users who use
flakes and have multiple nixos configurations with different
architectures. Because of https://github.com/NixOS/nix/issues/4265
this causes `nix flake check` to fail for them.

Note that users of nix-software-center that do not rely on the flake
will still need an IFD. I'm not aware of a better way (other than
upstreaming both in nixpkgs).

Ref: https://github.com/snowfallorg/nixos-appstream-data/pull/1